src/requirements.txt
requests
beautifulsoup4
pyyaml
tenacity
selenium
webdriver-manager